Bouge l'inscription dans son proipre composant
